Compile an R Markdown file to multiple outputs

This function takes an R Markdown file as input and generates three outputs: an HTML document, a PDF-like HTML document using the pagedown package, and a PDF document. If generating the PDF document fails (e.g., if Chrome is not installed), the function will still generate the PDF-like HTML document.

This function takes an R Markdown file as input and generates three outputs: an HTML document, a PDF-like HTML document using the pagedown package, and a PDF document. If generating the PDF document fails (e.g., if Chrome is not installed), the function will still generate the PDF-like HTML document.

compile_rmd_multiple_outputs(input_file, max_wait = 10)

compile_rmd_multiple_outputs(input_file, max_wait = 10)

Arguments

input_file

The path to the R Markdown file to be compiled.

max_wait

Max. waittime (minutes) until the pdf finishes printing. Default = 10

Value

Invisibly returns NULL. The side effect of this function is that it writes the compiled documents to disk (same folder as source).

Invisibly returns NULL. The side effect of this function is that it writes the compiled documents to disk (same folder as source).

Examples

if (FALSE) {
compile_rmd_multiple_outputs("path/to/your_file.Rmd")
}

if (FALSE) {
compile_rmd_multiple_outputs("path/to/your_file.Rmd")
}