Apply style to a given table

This function applies a particular style to a table. The table can be a data frame, a gtsummary table, or a knitr kable. Different processing steps are applied based on the class of the input table.

martys_table_style(
  tabl,
  alig. = c("c"),
  caption. = "",
  scale_down. = FALSE,
  fontsize. = 7,
  landscape. = TRUE
)

Arguments

tabl

The table to which the style should be applied.

alig.

The alignment to be applied to the table. Default is "c" for center.

caption.

The caption of the table. Default is the result of generate_tab_title().

scale_down.

Logical, if TRUE, scales down the table if it is wider than the text width. Default is FALSE.

fontsize.

The font size to be applied to the table. Default is 7.

landscape.

Logical, if TRUE, the table will be displayed in landscape format. Default is TRUE.

Value

A table with the specified style applied.

Examples

if (FALSE) {
data(mtcars)
martys_table_style(mtcars)
}